LayoutTests/fast/speechrecognition/permission-error.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test.js"></script>
<script>
description("Verify that error event is fired when permission is denied.");

if (window.testRunner) {
    jsTestIsAsync = true;
    testRunner.setIsSpeechRecognitionPermissionGranted(false);
}

shouldNotThrow("recognition = new SpeechRecognition()");
shouldNotThrow("recognition.start()");
recognition.onerror = (event) => {
    shouldBeEqualToString("event.error", "not-allowed");
    shouldBeEqualToString("event.message", "Permission check failed");

    finishJSTest();
}

</script>
</body>
</html>
